requirements.txt
requests
openpyxl
python-whois
flask
urllib3
